Fix chemicalish blasts not spawning particles
Add better usage for the rocket launcher in creative

Implement new missile models

Fix wrong config data type

Implement proper supportframe functionality

Actually use power for launches

+++ src/main/java/ballistix/common/settings/Constants.java
@@ -144,6 +144,6 @@ public class Constants {
 	public static boolean SHOULD_CACHE_EXPLOSIONS = true;
 
 	@IntValue(def = 1000)
-	public static double LASER_DESIGNATOR_RANGE = 1000;
+	public static int LASER_DESIGNATOR_RANGE = 1000;
 
 }
